99 dreamlet-edu-video 2/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "ffmpeg-python>=0.2.0",
    "numpy>=2.2.4",
    "openai>=1.72.0",
    "opencv-python-headless>=4.11.0.86",
    "opencv-python>=4.11.0.86",
    "pillow>=11.1.0",
    "python-docx>=1.1.2",
    "python-pptx>=1.0.2",
    "setuptools>=78.1.0",
    "streamlit>=1.44.1",
    "pdf2image>=1.17.0",
    "psutil>=7.0.0",
    "anthropic>=0.50.0",
    "requests>=2.32.3",
    "pandas>=2.2.3",
]
